+++ packages/manager/.changeset/pr-13314-added-1769167928025.md
@@ -1,0 +1,5 @@
+---
+"@linode/manager": Added
+---
+
+IAM Delegation: empty state for user delegations ([#13314](https://github.com/linode/manager/pull/13314))
